Wait statements inserted with DEBUG Commands (#1)
* Wait statements inserted with DEBUG Commands

Currently the wait statement waits a total of 5 minutes writing a DEBUG statement out to CloudWatch. It should be noted that it can also be delivered potentially to a 3rd party's logger which may help the attacker retain some type of out of band messaging. The wait statement has been helpful in stopping / starting the UserDataSwap with automatic build tools.

Currently has not been fully tested with Autoscaling groups.

* Formatting fixes in main.go

Co-authored-by: Ryan Gerstenkorn <me@ryanjarv.sh>